Upload refactoring: remove old js
