+++ public/js/cat_source/es6/components/common/ApplicationWrapper/ApplicationWrapperContext.js
@@ -1,0 +1,3 @@
+import {createContext} from 'react'
+
+export const ApplicationWrapperContext = createContext({})
