+++ public/js/cat_source/es6/components/createProject/TmGlossarySelect.js
@@ -3,7 +3,7 @@ import Check from '../../../../../img/icons/Check'
 import {Select} from '../common/Select'
 import {CreateProjectContext} from './CreateProjectContext'
 import {orderTmKeys} from '../settingsPanel/Contents/TranslationMemoryGlossaryTab'
-import {ApplicationWrapperContext} from '../common/ApplicationWrapper'
+import {ApplicationWrapperContext} from '../common/ApplicationWrapper/ApplicationWrapperContext'
 
 export const TmGlossarySelect = () => {
   const {
