+++ public/js/cat_source/es6/components/header/cattol/Header.js
@@ -13,7 +13,7 @@ import {SegmentsFilterButton} from './SegmentsFilterButton'
 import {SettingsButton} from './SettingsButton'
 import {ActionMenu} from '../ActionMenu'
 import {UserMenu} from '../UserMenu'
-import {ApplicationWrapperContext} from '../../common/ApplicationWrapper'
+import {ApplicationWrapperContext} from '../../common/ApplicationWrapper/ApplicationWrapperContext'
 
 export const Header = ({
   jid,
