+++ public/js/cat_source/es6/components/modals/PreferencesModal.js
@@ -17,7 +17,7 @@ import {
 import IconClose from '../icons/IconClose'
 import UserStore from '../../stores/UserStore'
 import {getUserData} from '../../api/getUserData'
-import {ApplicationWrapperContext} from '../common/ApplicationWrapper'
+import {ApplicationWrapperContext} from '../common/ApplicationWrapper/ApplicationWrapperContext'
 import ModalsActions from '../../actions/ModalsActions'
 import useAuth from '../../hooks/useAuth'
