+++ public/js/cat_source/es6/components/settingsPanel/Contents/EditorSettingsTab/CrossLanguagesMatches.js
@@ -2,7 +2,7 @@ import React, {useContext, useEffect, useRef, useState} from 'react'
 import {Select} from '../../../common/Select'
 import SegmentActions from '../../../../actions/SegmentActions'
 import ApplicationStore from '../../../../stores/ApplicationStore'
-import {ApplicationWrapperContext} from '../../../common/ApplicationWrapper'
+import {ApplicationWrapperContext} from '../../../common/ApplicationWrapper/ApplicationWrapperContext'
 import {isEqual} from 'lodash'
 import SegmentStore from '../../../../stores/SegmentStore'
 import {METADATA_KEY} from '../../../../constants/Constants'
