+++ public/js/cat_source/es6/components/settingsPanel/Contents/EditorSettingsTab/SpacePlaceholder.js
@@ -2,7 +2,7 @@ import React, {useContext, useState} from 'react'
 import Switch from '../../../common/Switch'
 import SegmentActions from '../../../../actions/SegmentActions'
 import {setTagSignatureMiddleware} from '../../../segments/utils/DraftMatecatUtils/tagModel'
-import {ApplicationWrapperContext} from '../../../common/ApplicationWrapper'
+import {ApplicationWrapperContext} from '../../../common/ApplicationWrapper/ApplicationWrapperContext'
 
 const METADATA_KEY = 'show_whitespace'
