+++ public/js/cat_source/es6/hooks/useSegmentsLoader.js
@@ -4,7 +4,7 @@ import {getSegments} from '../api/getSegments'
 import SegmentActions from '../actions/SegmentActions'
 import SegmentStore from '../stores/SegmentStore'
 import CommonUtils from '../utils/commonUtils'
-import {ApplicationWrapperContext} from '../components/common/ApplicationWrapper'
+import {ApplicationWrapperContext} from '../components/common/ApplicationWrapper/ApplicationWrapperContext'
 
 const INIT_NUM_SEGMENTS = 40
 const MORE_NUM_SEGMENTS = 25
