+++ public/js/cat_source/es6/pages/XliffToTarget.js
@@ -1,7 +1,7 @@
 import React, {useContext} from 'react'
 import usePortal from '../hooks/usePortal'
 import Header from '../components/header/Header'
-import {ApplicationWrapperContext} from '../components/common/ApplicationWrapper'
+import {ApplicationWrapperContext} from '../components/common/ApplicationWrapper/ApplicationWrapperContext'
 import Footer from '../components/footer/Footer'
 import SocketListener from '../sse/SocketListener'
 import {mountPage} from './mountPage'
