+++ public/js/cat_source/es6/sse/SocketListener.js
@@ -5,7 +5,7 @@ import SegmentActions from '../actions/SegmentActions'
 import SegmentStore from '../stores/SegmentStore'
 import CommonUtils from '../utils/commonUtils'
 import CommentsActions from '../actions/CommentsActions'
-import {ApplicationWrapperContext} from '../components/common/ApplicationWrapper'
+import {ApplicationWrapperContext} from '../components/common/ApplicationWrapper/ApplicationWrapperContext'
 import UserActions from '../actions/UserActions'
 import {v4 as uuidV4} from 'uuid'
 import Cookies from 'js-cookie'
